requirements.txt
groq==0.4.2
gradio==4.19.2
python-dotenv==1.0.0 